requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.15.2
sqlparse==0.5.1
django-cors-headers==4.5.0


